backend/modules/lunar/requirements.txt
fastapi
uvicorn
lunar-python
opencc-python-reimplemented
